initial commit
I've already had an "Exercises A" repository (predominantly for code)
for more than a decade. There's never been an occasion to make
"Exercises B". (It wasn't like I was going to run out of paper for the
first repository.) But if I'm going to be LaTeXing up more math
exercises instead of doing them all on paper, I don't feel like that
belongs in the same repo as code exercises, so maybe this can be
Exercises B?