maybe skip the commutative diagram
I was getting mysterious pdflatex compile errors (`Fatal format file
error; I'm stymied`) that actually didn't seem causally related to the
diagram (in that you can take the diagram out and it's still broken),
that were fixable with the magic command `sudo fmtutil-sys --all`, but
it then broke again later.

Removing the diagram is a embarrassingly superstitious response (if
that were really the problem, you'd expect the compilation error to
deterministically be due to the presence of the diagram code), but the
temporal proximity makes me suspicious (by the explaining-away effect,
it would be a weird coincidence that I see this obscure error on the
same day that I try a new LaTeX feature) and it's just not that
important.